+++ Flow/src/utopia/flow/generic/ModelConvertible.scala
@@ -2,14 +2,14 @@ package utopia.flow.generic
 
 import utopia.flow.datastructure.immutable.Model
 import utopia.flow.datastructure.immutable.Constant
-import utopia.flow.parse.JSONConvertible
+import utopia.flow.parse.JsonConvertible
 
 /**
  * ModelConvertible instances can be represented as model data when necessary
  * @author Mikko Hilpinen
  * @since 23.6.2017
  */
-trait ModelConvertible extends JSONConvertible
+trait ModelConvertible extends JsonConvertible
 {
     // ABSTRACT METHODS & PROPERTIES    -----------------
     
@@ -21,7 +21,7 @@ trait ModelConvertible extends JSONConvertible
     
     // IMPLEMENTED    ---------------------------
     
-    override def toJSON = toModel.toJSON
+    override def toJson = toModel.toJson
     
     override def toString = toModel.toString
 }