+++ Flow/src/utopia/flow/parse/BasicJsonValueConverter.scala
@@ -5,11 +5,11 @@ import utopia.flow.datastructure.immutable.Value
 import utopia.flow.generic.{BooleanType, DataType, DoubleType, FloatType, InstantType, IntType, LocalDateType, LongType, ModelType, StringType, VectorType}
 
 /**
- * This JSON value writer is able to write instances of basic data types into JSON
+ * This json value writer is able to write instances of basic data types into json
  * @author Mikko Hilpinen
  * @since 17.12.2016
  */
-object BasicJSONValueConverter extends ValueConverter[String]
+object BasicJsonValueConverter extends ValueConverter[String]
 {
     override val supportedTypes = HashSet(StringType, VectorType, ModelType, IntType, DoubleType,
             FloatType, LongType, BooleanType, InstantType)
@@ -20,8 +20,8 @@ object BasicJSONValueConverter extends ValueConverter[String]
         {
             case StringType => "\"" + value.getString.replace("\"", "'")
                 .replace("\\", "\\\\") + "\""
-            case VectorType => s"[${value.getVector.map { _.toJSON }.mkString(", ")}]"
-            case ModelType => value.getModel.toJSON
+            case VectorType => s"[${value.getVector.map { _.toJson }.mkString(", ")}]"
+            case ModelType => value.getModel.toJson
             // Handles instant type separately to format it correctly
             case InstantType => "\"" + value.getString + "\""
             // Same treatment is given to dates