+++ Vault/test/utopia/vault/test/Person.scala
@@ -6,9 +6,9 @@ import utopia.flow.datastructure.immutable
 import utopia.flow.datastructure.immutable.Constant
 import utopia.flow.generic.ValueConversions._
 import utopia.vault.model.immutable.StorableWithFactory
-import utopia.vault.nosql.factory.StorableFactoryWithValidation
+import utopia.vault.nosql.factory.FromValidatedRowModelFactory
 
-object Person extends StorableFactoryWithValidation[Person]
+object Person extends FromValidatedRowModelFactory[Person]
 {
     // ATTRIBUTES    ----------------